Changed Profile Scenes * Changed ShowProfileActivity to ShowProfileView * Created UserSingleton * Implemented UserSingleton in ShowProfileWorker
